Code/requirements.txt
opencv-python==4.5.4.58
easyocr==1.4
imutils==0.5.4
mysql-connector-python==8.0.27
numpy==1.21.3
pytesseract==0.3.8
ultralytics==0.0.7
tk==0.1.0
reportlab==3.6.1
